analyzer: add intermediate granularity
This granularity allows finding which packages exist along a path in the
callgraph from a function in a queried package to a capability.

This allows users to see when a change in their dependencies results in
a capability being reached via a new package, even if the top-level
function already had that capability through another path.