all: allow importing of external callgraph and capability information
This separates the work of cmd/capslock into two phases.  In the first
phase, the specified packages are loaded, a callgraph is computed, and
functions are labelled with capabilities.  This data is stored in a
language-independent format.  In the second phase, the callgraph and
capabilities are analyzed to produce the required output.

We add a flag to allow importing the language-independent format from a
file, so that external programs can be used to produce callgraphs for
other languages.